Extending simulation to include enemy interactions
This is for the "realistic drops" mode. It also introduces a 10-second
timer after killing a knight; the shape must be interacted with before
the timer runs out or it will disable itself.

And when hints are enabled, some extra hints give direction, including
highlighting the invalid shapes present for a statue.